% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models_by_treatment.R
\name{compare_models_by_treatment}
\alias{compare_models_by_treatment}
\title{Compare models by treatment}
\usage{
compare_models_by_treatment(...)
}
\arguments{
\item{...}{Fitted model objects.}
}
\value{
A data frame.
}
\description{
Calculates model performance separately
for each treatment.
}
