% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_heads.R
\name{exclude_heads}
\alias{exclude_heads}
\title{Exclude Problematic ANKOM Heads}
\usage{
exclude_heads(data, heads, reason = NULL)
}
\arguments{
\item{data}{A \code{rumen_gp} object.}

\item{heads}{Character vector of bottle
identifiers to remove.}

\item{reason}{Character vector containing the
reason for each exclusion.}
}
\value{
A filtered \code{rumen_gp} object.
}
\description{
Removes one or more bottles from a
\code{rumen_gp} object while recording
exclusion information.
}
\details{
This function is useful for excluding:

\itemize{
\item Leaking bottles
\item Sensor failures
\item Damaged bottles
\item Biologically implausible observations
\item Other quality-control issues
}

Exclusion information is retained to support
transparent reporting and reproducible analyses.
}
\examples{

files <- example_data()

raw_data <- read_ankom(
  files$ankom
)

metadata <- read_metadata(
  files$metadata
)

gp <- process_ankom(
  raw_data,
  metadata,
  headspace_ml = 210,
  temperature_c = 39
)

gp_filtered <- exclude_heads(
  data = gp,
  heads = c(
    "12",
    "18"
  ),
  reason = c(
    "Bottle leak",
    "Sensor malfunction"
  )
)

gp_filtered

}
\seealso{
\code{\link{validate_ankom}},
\code{\link{flag_model}},
\code{\link{process_ankom}}
}
