% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_brody.R
\name{fit_brody}
\alias{fit_brody}
\title{Fit Brody model}
\usage{
fit_brody(data, start = NULL)
}
\arguments{
\item{data}{A rumen_gp object.}

\item{start}{Optional list of starting values.
May contain any of:
A, b, and k.}
}
\value{
A brody_fit object.
}
\description{
Fits the Brody gas production model
to each ANKOM bottle.
}
