% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gompertz.R
\name{fit_gompertz}
\alias{fit_gompertz}
\title{Fit Gompertz model}
\usage{
fit_gompertz(data, start = NULL)
}
\arguments{
\item{data}{A rumen_gp object.}

\item{start}{Optional list of starting values.
May contain any of:
A, mu, and lambda.}
}
\value{
A list containing:
\itemize{
\item parameters
\item diagnostics
\item predictions
}
}
\description{
Fits the Zwietering-modified Gompertz model
to each ANKOM bottle.
}
