% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lel.R
\name{fit_lel}
\alias{fit_lel}
\title{Fit Logistic-Exponential model (LEL)}
\usage{
fit_lel(data, start = NULL)
}
\arguments{
\item{data}{A rumen_gp object.}

\item{start}{Optional list of starting values.
May contain any of:
A, k, d, and lambda.}
}
\value{
A lel_fit object.
}
\description{
Fits the Logistic-Exponential model with
an explicit lag phase.
}
