% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_richards.R
\name{fit_richards}
\alias{fit_richards}
\title{Fit Richards model}
\usage{
fit_richards(data)
}
\arguments{
\item{data}{A rumen_gp object.}
}
\value{
A richards_fit object.
}
\description{
Fits the Richards gas production model
to each ANKOM bottle.
}
