% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_ankom_time.R
\name{parse_ankom_time}
\alias{parse_ankom_time}
\title{Parse ANKOM Timestamps}
\usage{
parse_ankom_time(time_raw)
}
\arguments{
\item{time_raw}{Character vector containing
ANKOM timestamps.}
}
\value{
A numeric vector containing elapsed
incubation time in hours.
}
\description{
Converts ANKOM RF timestamps into elapsed
incubation time expressed in hours.
}
\details{
ANKOM RF systems record measurements using
timestamps. This function converts those
timestamps into elapsed incubation time,
measured relative to the first observation.

The resulting values are used throughout
rumenGP for:

\itemize{
\item Data processing
\item Model fitting
\item Visualization
\item Model comparison
}

In most workflows, this function is called
automatically by \code{process_ankom()} and
does not need to be used directly.
}
\examples{

timestamps <- c(
  "2024-01-01 08:00:00",
  "2024-01-01 12:00:00",
  "2024-01-01 20:00:00"
)

parse_ankom_time(
  timestamps
)

# Typical workflow
files <- example_data()

raw_data <- read_ankom(
  files$ankom
)

metadata <- read_metadata(
  files$metadata
)

gp <- process_ankom(
  raw_data,
  metadata,
  headspace_ml = 210,
  temperature_c = 39
)

head(
  gp$Time_h
)

}
\seealso{
\code{\link{read_ankom}},
\code{\link{process_ankom}},
\code{\link{example_data}}
}
