% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_all_gompertz_fits.R
\name{plot_all_gompertz_fits}
\alias{plot_all_gompertz_fits}
\title{Plot all Gompertz fits}
\usage{
plot_all_gompertz_fits(fit)
}
\arguments{
\item{fit}{A gompertz_fit object.}
}
\value{
A ggplot object.
}
\description{
Displays observed and predicted gas
production for all bottles.
}
