% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gompertz_fit.R
\name{plot_gompertz_fit}
\alias{plot_gompertz_fit}
\title{Plot Gompertz fit}
\usage{
plot_gompertz_fit(fit, head)
}
\arguments{
\item{fit}{A gompertz_fit object.}

\item{head}{Head identifier.}
}
\value{
A ggplot object.
}
\description{
Plots observed and predicted gas production
for an individual bottle.
}
