% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gp.R
\name{plot_gp}
\alias{plot_gp}
\title{Plot raw gas production curve}
\usage{
plot_gp(data, head)
}
\arguments{
\item{data}{A rumen_gp object.}

\item{head}{Head identifier.}
}
\value{
A ggplot object.
}
\description{
Plots observed gas production for a bottle.
}
