% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_comparison.R
\name{plot_model_comparison}
\alias{plot_model_comparison}
\title{Compare model fits}
\usage{
plot_model_comparison(..., head)
}
\arguments{
\item{...}{Fitted model objects.}

\item{head}{Head identifier.}
}
\value{
A ggplot object.
}
\description{
Plots observed data together with
multiple model predictions.
}
