% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_comparison_all.R
\name{plot_model_comparison_all}
\alias{plot_model_comparison_all}
\title{Compare models for all bottles}
\usage{
plot_model_comparison_all(...)
}
\arguments{
\item{...}{Fitted model objects.}
}
\value{
A ggplot object.
}
\description{
Displays observed and predicted values for
multiple fitted models across all bottles.
}
