% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residual_comparison.R
\name{plot_residual_comparison}
\alias{plot_residual_comparison}
\title{Compare residuals across models}
\usage{
plot_residual_comparison(..., head)
}
\arguments{
\item{...}{Fitted model objects.}

\item{head}{Head identifier.}
}
\value{
A ggplot object.
}
\description{
Compare residuals across models
}
