% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressure_to_volume.R
\name{pressure_to_volume}
\alias{pressure_to_volume}
\title{Convert Pressure to Gas Volume}
\usage{
pressure_to_volume(
  pressure,
  pressure_unit = c("psi", "kpa"),
  headspace_volume,
  headspace_unit = c("mL", "L"),
  temperature = 39
)
}
\arguments{
\item{pressure}{Numeric pressure values.}

\item{pressure_unit}{Pressure unit.
One of:
\itemize{
\item \code{"psi"}
\item \code{"kpa"}
}}

\item{headspace_volume}{Headspace volume.}

\item{headspace_unit}{Headspace volume unit.
One of:
\itemize{
\item \code{"mL"}
\item \code{"L"}
}}

\item{temperature}{Incubation temperature
in degrees Celsius.}
}
\value{
A numeric vector containing
estimated gas volumes in mL.
}
\description{
Converts pressure measurements to estimated
gas volumes using the ideal gas law.
}
\details{
The function supports pressure measurements
in PSI or kPa and headspace volumes in mL or L.
\subsection{Equation}{

Gas volume is estimated using the ideal gas law:

\deqn{
PV = nRT
}

where:

\itemize{
\item \eqn{P} is pressure
\item \eqn{V} is headspace volume
\item \eqn{n} is gas moles
\item \eqn{R} is the gas constant
\item \eqn{T} is absolute temperature
}

Estimated gas moles are converted to an
equivalent gas volume.
}
}
\examples{
# Convert PSI measurements
pressure_to_volume(
  pressure = c(
    0.5,
    1.0,
    1.5
  ),
  pressure_unit = "psi",
  headspace_volume = 60,
  headspace_unit = "mL",
  temperature = 39
)

# Convert kPa measurements
pressure_to_volume(
  pressure = c(
    5,
    10,
    15
  ),
  pressure_unit = "kpa",
  headspace_volume = 0.06,
  headspace_unit = "L",
  temperature = 39
)

}
