% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ankom.R
\name{process_ankom}
\alias{process_ankom}
\title{Process ANKOM RF data}
\usage{
process_ankom(
  raw_data,
  metadata = NULL,
  headspace_ml = 210,
  temperature_c = 39
)
}
\arguments{
\item{raw_data}{Raw ANKOM table.}

\item{metadata}{Metadata table.}

\item{headspace_ml}{Headspace volume.}

\item{temperature_c}{Incubation temperature.}
}
\value{
Processed data frame.
}
\description{
Converts raw ANKOM RF output into a standardized dataset.
}
