% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_models_by_treatment.R
\name{rank_models_by_treatment}
\alias{rank_models_by_treatment}
\title{Rank models within each treatment}
\usage{
rank_models_by_treatment(comparison)
}
\arguments{
\item{comparison}{Output from compare_models_by_treatment().}
}
\value{
Ranked data frame.
}
\description{
Rank models within each treatment
}
