% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{read_metadata}
\alias{read_metadata}
\title{Import Metadata}
\usage{
read_metadata(file, sheet = "metadata")
}
\arguments{
\item{file}{Path to a metadata file.}

\item{sheet}{Sheet name containing metadata.}
}
\value{
A data frame containing experimental
metadata.
}
\description{
Reads metadata associated with an ANKOM RF
experiment.
}
\details{
Metadata are used to identify bottles,
treatments, replicates, and other experimental
information required for downstream analyses.

This function is typically used together with:

\itemize{
\item \code{read_ankom()}
\item \code{process_ankom()}
}

as part of the standard ANKOM workflow.
}
\examples{

files <- example_data()

metadata <- read_metadata(
  files$metadata
)

head(
  metadata
)

# Typical workflow
raw_data <- read_ankom(
  files$ankom
)

gp <- process_ankom(
  raw_data,
  metadata,
  headspace_ml = 210,
  temperature_c = 39
)

head(
  gp
)

}
\seealso{
\code{\link{read_ankom}},
\code{\link{process_ankom}},
\code{\link{validate_metadata}},
\code{\link{example_data}}
}
