% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.custom_fit.R
\name{summary.custom_fit}
\alias{summary.custom_fit}
\title{Summary of custom model fits}
\usage{
\method{summary}{custom_fit}(object, ...)
}
\arguments{
\item{object}{A custom_fit object.}

\item{...}{Not used.}
}
\value{
Invisibly returns the input object.
}
\description{
Summarizes a custom_fit object.
}
