% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.dual_logistic_fit.R
\name{summary.dual_logistic_fit}
\alias{summary.dual_logistic_fit}
\title{Summary of dual logistic fits}
\usage{
\method{summary}{dual_logistic_fit}(object, ...)
}
\arguments{
\item{object}{A dual_logistic_fit object.}

\item{...}{Additional arguments.}
}
\description{
Summary of dual logistic fits
}
