% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.gompertz_fit.R
\name{summary.gompertz_fit}
\alias{summary.gompertz_fit}
\title{Summary of Gompertz fits}
\usage{
\method{summary}{gompertz_fit}(object, ...)
}
\arguments{
\item{object}{A gompertz_fit object.}

\item{...}{Additional arguments.}
}
\description{
Summarizes model fitting results.
}
