% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.groot_fit.R
\name{summary.groot_fit}
\alias{summary.groot_fit}
\title{Summary of Groot fits}
\usage{
\method{summary}{groot_fit}(object, ...)
}
\arguments{
\item{object}{A groot_fit object.}

\item{...}{Additional arguments.}
}
\description{
Summary of Groot fits
}
