% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.le0_fit.R
\name{summary.le0_fit}
\alias{summary.le0_fit}
\title{Summary of LE0 fits}
\usage{
\method{summary}{le0_fit}(object, ...)
}
\arguments{
\item{object}{A le0_fit object.}

\item{...}{Additional arguments.}
}
\description{
Summary of LE0 fits
}
