% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.logistic_fit.R
\name{summary.logistic_fit}
\alias{summary.logistic_fit}
\title{Summary of Logistic fits}
\usage{
\method{summary}{logistic_fit}(object, ...)
}
\arguments{
\item{object}{A logistic_fit object.}

\item{...}{Additional arguments.}
}
\description{
Summarizes Logistic model fitting results.
}
