% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.richards_fit.R
\name{summary.richards_fit}
\alias{summary.richards_fit}
\title{Summary of Richards fits}
\usage{
\method{summary}{richards_fit}(object, ...)
}
\arguments{
\item{object}{A richards_fit object.}

\item{...}{Additional arguments.}
}
\description{
Summary of Richards fits
}
