% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_metadata.R
\name{validate_metadata}
\alias{validate_metadata}
\title{Validate metadata}
\usage{
validate_metadata(metadata)
}
\arguments{
\item{metadata}{Metadata table.}
}
\value{
The validated metadata table.
}
\description{
Validates ANKOM metadata before analysis.
}
\details{
Required columns:
\itemize{
\item Head
\item Treatment
\item Rep
}
}
