% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_model_by_treatment.R
\name{best_model_by_treatment}
\alias{best_model_by_treatment}
\title{Identify the best model for each treatment}
\usage{
best_model_by_treatment(ranked_comparison)
}
\arguments{
\item{ranked_comparison}{Output from rank_models_by_treatment().}
}
\value{
A data frame containing the best model for each treatment.
}
\description{
Returns the top-ranked model within each treatment.
}
