% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_data}
\alias{example_data}
\title{Example ANKOM data}
\usage{
example_data()
}
\value{
A list containing file paths.
}
\description{
Returns paths to example files included
with the package.
}
