% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_heads.R
\name{exclude_heads}
\alias{exclude_heads}
\title{Exclude problematic ANKOM heads}
\usage{
exclude_heads(data, heads, reason = NULL)
}
\arguments{
\item{data}{A rumen_gp object.}

\item{heads}{Character vector of heads to remove.}

\item{reason}{Character vector of exclusion reasons.}
}
\value{
A filtered rumen_gp object.
}
\description{
Removes specified bottles while recording
the exclusion information.
}
