% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_brody.R
\name{fit_brody}
\alias{fit_brody}
\title{Fit Brody model}
\usage{
fit_brody(data)
}
\arguments{
\item{data}{A rumen_gp object.}
}
\value{
A brody_fit object.
}
\description{
Fits the Brody gas production model
to each ANKOM bottle.
}
