% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dual_logistic.R
\name{fit_dual_logistic}
\alias{fit_dual_logistic}
\title{Fit dual-pool logistic model}
\usage{
fit_dual_logistic(data)
}
\arguments{
\item{data}{A rumen_gp object.}
}
\value{
A dual_logistic_fit object.
}
\description{
Fits a dual-pool logistic model representing
rapidly and slowly degradable fractions.
}
