% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_exp0.R
\name{fit_exp0}
\alias{fit_exp0}
\title{Fit exponential model without lag (EXP0)}
\usage{
fit_exp0(data)
}
\arguments{
\item{data}{A rumen_gp object.}
}
\value{
An exp0_fit object.
}
\description{
Fits the exponential gas production model
without an explicit lag phase.
}
