% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_expl.R
\name{fit_expl}
\alias{fit_expl}
\title{Fit exponential model with lag (EXPL)}
\usage{
fit_expl(data, start = NULL)
}
\arguments{
\item{data}{A rumen_gp object.}

\item{start}{Optional list of starting values.
May contain any of:
Vf, k, and lambda.}
}
\value{
An expl_fit object.
}
\description{
Fits the exponential gas production model
with an explicit lag phase.
}
