% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_expl.R
\name{fit_expl}
\alias{fit_expl}
\title{Fit exponential model with lag (EXPL)}
\usage{
fit_expl(data)
}
\arguments{
\item{data}{A rumen_gp object.}
}
\value{
An expl_fit object.
}
\description{
Fits the exponential gas production model
with an explicit lag phase.
}
