% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gompertz.R
\name{fit_gompertz}
\alias{fit_gompertz}
\title{Fit Gompertz model}
\usage{
fit_gompertz(data)
}
\arguments{
\item{data}{A rumen_gp object.}
}
\value{
A list containing parameter estimates,
diagnostics and predictions.
}
\description{
Fits the Zwietering-modified Gompertz model
to each ANKOM bottle.
}
