% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_groot.R
\name{fit_groot}
\alias{fit_groot}
\title{Fit Groot model}
\usage{
fit_groot(data, start = NULL)
}
\arguments{
\item{data}{A rumen_gp object.}

\item{start}{Optional list of starting values.
May contain any of:
VF, b, and k.}
}
\value{
A groot_fit object.
}
\description{
Fits the Groot gas production model
to each ANKOM bottle.
}
