% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_le0.R
\name{fit_le0}
\alias{fit_le0}
\title{Fit Logistic-Exponential model (LE0)}
\usage{
fit_le0(data)
}
\arguments{
\item{data}{A rumen_gp object.}
}
\value{
A le0_fit object.
}
\description{
Fits the Logistic-Exponential model without
an explicit lag phase.
}
