% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_le0.R
\name{fit_le0}
\alias{fit_le0}
\title{Fit Logistic-Exponential model (LE0)}
\usage{
fit_le0(data, start = NULL)
}
\arguments{
\item{data}{A rumen_gp object.}

\item{start}{Optional list of starting values.
May contain any of:
A, k, and d.}
}
\value{
A le0_fit object.
}
\description{
Fits the Logistic-Exponential model without
an explicit lag phase.
}
