% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lel.R
\name{fit_lel}
\alias{fit_lel}
\title{Fit Logistic-Exponential model (LEL)}
\usage{
fit_lel(data)
}
\arguments{
\item{data}{A rumen_gp object.}
}
\value{
A lel_fit object.
}
\description{
Fits the Logistic-Exponential model with
an explicit lag phase.
}
