% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_logistic.R
\name{fit_logistic}
\alias{fit_logistic}
\title{Fit Logistic model}
\usage{
fit_logistic(data)
}
\arguments{
\item{data}{A rumen_gp object.}
}
\value{
A logistic_fit object.
}
\description{
Fits a Logistic gas-production model
to each ANKOM bottle.
}
