% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mitscherlich.R
\name{fit_mitscherlich}
\alias{fit_mitscherlich}
\title{Fit Mitscherlich model}
\usage{
fit_mitscherlich(data, start = NULL)
}
\arguments{
\item{data}{A rumen_gp object.}

\item{start}{Optional list of starting values.
May contain any of:
A, k, d, and lambda.}
}
\value{
A mitscherlich_fit object.
}
\description{
Fits the Mitscherlich gas-production model
to each ANKOM bottle.
}
