% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mm.R
\name{fit_mm}
\alias{fit_mm}
\title{Fit Michaelis-Menten model}
\usage{
fit_mm(data, start = NULL)
}
\arguments{
\item{data}{A rumen_gp object.}

\item{start}{Optional list of starting values.
May contain any of:
A, K, and c.}
}
\value{
A mm_fit object.
}
\description{
Fits the Michaelis-Menten gas production model
to each ANKOM bottle.
}
