% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mm.R
\name{fit_mm}
\alias{fit_mm}
\title{Fit Michaelis-Menten model}
\usage{
fit_mm(data)
}
\arguments{
\item{data}{A rumen_gp object.}
}
\value{
A mm_fit object.
}
\description{
Fits the Michaelis-Menten gas production model
to each ANKOM bottle.
}
