% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_orskov.R
\name{fit_orskov}
\alias{fit_orskov}
\title{Fit Orskov and McDonald model}
\usage{
fit_orskov(data, start = NULL)
}
\arguments{
\item{data}{A rumen_gp object.}

\item{start}{Optional list of starting values.
May contain any of:
VF, b, and k.}
}
\value{
An orskov_fit object.
}
\description{
Fits the Orskov and McDonald gas
production model to each ANKOM bottle.
}
