% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_orskov.R
\name{fit_orskov}
\alias{fit_orskov}
\title{Fit Ørskov and McDonald model}
\usage{
fit_orskov(data)
}
\arguments{
\item{data}{A rumen_gp object.}
}
\value{
An orskov_fit object.
}
\description{
Fits the Ørskov and McDonald gas
production model to each ANKOM bottle.
}
