% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_win_frequency.R
\name{model_win_frequency}
\alias{model_win_frequency}
\title{Model win frequency}
\usage{
model_win_frequency(best_models)
}
\arguments{
\item{best_models}{Output from
best_model_by_treatment().}
}
\value{
A data frame.
}
\description{
Summarizes how often each model is the
best-performing model across treatments.
}
