% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_ankom_time.R
\name{parse_ankom_time}
\alias{parse_ankom_time}
\title{Parse ANKOM timestamps}
\usage{
parse_ankom_time(time_raw)
}
\arguments{
\item{time_raw}{Character vector containing ANKOM timestamps.}
}
\value{
Numeric vector of elapsed incubation time in hours.
}
\description{
Converts ANKOM RF timestamps into elapsed incubation time (hours).
}
