% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_all_fits.R
\name{plot_all_fits}
\alias{plot_all_fits}
\title{Plot all fitted curves}
\usage{
plot_all_fits(fit)
}
\arguments{
\item{fit}{A fitted model object.}
}
\value{
A ggplot object.
}
\description{
Displays observed and predicted values
for all bottles.
}
