% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostics.R
\name{plot_diagnostics}
\alias{plot_diagnostics}
\title{Plot diagnostic summaries}
\usage{
plot_diagnostics(fit)
}
\arguments{
\item{fit}{A fitted model object.}
}
\value{
A list of ggplots.
}
\description{
Diagnostic histograms for a fitted model.
}
