% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fit.R
\name{plot_fit}
\alias{plot_fit}
\title{Plot fitted model}
\usage{
plot_fit(fit, head)
}
\arguments{
\item{fit}{A fitted model object.}

\item{head}{Head identifier.}
}
\value{
A ggplot object.
}
\description{
Plots observed and predicted gas production
for an individual bottle.
}
