% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gp.R
\name{plot_gp}
\alias{plot_gp}
\title{Plot Raw Gas Production Curve}
\usage{
plot_gp(data, head)
}
\arguments{
\item{data}{A \code{rumen_gp} object.}

\item{head}{Bottle identifier to plot.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plots observed gas production measurements
for an individual bottle.
}
\details{
This visualization displays the raw gas
production profile prior to model fitting and
is useful for:

\itemize{
\item Inspecting fermentation dynamics
\item Identifying unusual observations
\item Evaluating data quality
\item Comparing individual bottle profiles
}
}
\examples{

files <- example_data()

raw_data <- read_ankom(
  files$ankom
)

metadata <- read_metadata(
  files$metadata
)

gp <- process_ankom(
  raw_data,
  metadata,
  headspace_ml = 210,
  temperature_c = 39
)

plot_gp(
  gp,
  head = 1
)

}
\seealso{
\code{\link{plot_fit}},
\code{\link{plot_residuals}},
\code{\link{process_ankom}}
}
