% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_comparison_treatment.R
\name{plot_model_comparison_treatment}
\alias{plot_model_comparison_treatment}
\title{Compare models for a treatment}
\usage{
plot_model_comparison_treatment(..., treatment)
}
\arguments{
\item{...}{Fitted model objects.}

\item{treatment}{Treatment name.}
}
\value{
A ggplot object.
}
\description{
Displays observed and predicted values
for multiple models across all replicates
of a selected treatment.
}
