% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_performance.R
\name{plot_model_performance}
\alias{plot_model_performance}
\title{Plot model performance}
\usage{
plot_model_performance(comparison)
}
\arguments{
\item{comparison}{Output from compare_models().}
}
\value{
A ggplot object.
}
\description{
Visualizes model performance metrics.
}
