% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_rankings.R
\name{plot_model_rankings}
\alias{plot_model_rankings}
\title{Plot model rankings}
\usage{
plot_model_rankings(ranking)
}
\arguments{
\item{ranking}{Output from
rank_models().}
}
\value{
A ggplot object.
}
\description{
Visualizes model ranks across
performance metrics.
}
