% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residual_comparison.R
\name{plot_residual_comparison}
\alias{plot_residual_comparison}
\title{Compare Residuals Across Models}
\usage{
plot_residual_comparison(..., head)
}
\arguments{
\item{...}{Fitted model objects.}

\item{head}{Head identifier.}
}
\value{
A \code{ggplot2} object.
}
\description{
Displays residuals from multiple fitted models
for a selected bottle.
}
\details{
Residuals are calculated as:

\deqn{
Observed - Predicted
}

and can be used to evaluate:

\itemize{
\item Model bias
\item Systematic prediction errors
\item Heteroscedasticity
\item Relative model performance
}

Models with residuals that are randomly
distributed around zero are generally
preferred over models showing systematic
patterns.
}
\examples{

files <- example_data()

raw_data <- read_ankom(
  files$ankom
)

metadata <- read_metadata(
  files$metadata
)

gp <- process_ankom(
  raw_data,
  metadata,
  headspace_ml = 210,
  temperature_c = 39
)

groot_fit <- fit_groot(
  gp
)

gompertz_fit <- fit_gompertz(
  gp
)

plot_residual_comparison(
  Groot = groot_fit,
  Gompertz = gompertz_fit,
  head = 1
)

}
\seealso{
\code{\link{plot_residuals}},
\code{\link{plot_model_comparison}},
\code{\link{compare_models}},
\code{\link{fit_groot}},
\code{\link{fit_gompertz}}
}
