% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residuals.R
\name{plot_residuals}
\alias{plot_residuals}
\title{Plot model residuals}
\usage{
plot_residuals(fit, head = NULL)
}
\arguments{
\item{fit}{A fitted model object containing a
predictions element.}

\item{head}{Optional Head identifier.
If omitted and only one bottle is present,
that bottle is plotted automatically.}
}
\value{
A ggplot object.
}
\description{
Plots residuals for an individual bottle.
}
