% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residuals.R
\name{plot_residuals}
\alias{plot_residuals}
\title{Plot residuals from Gompertz fit}
\usage{
plot_residuals(fit, head)
}
\arguments{
\item{fit}{A gompertz_fit object.}

\item{head}{Head identifier.}
}
\value{
A ggplot object.
}
\description{
Plot residuals from Gompertz fit
}
