% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_treatment_mean.R
\name{plot_treatment_mean}
\alias{plot_treatment_mean}
\title{Plot treatment means across models}
\usage{
plot_treatment_mean(..., treatment)
}
\arguments{
\item{...}{Fitted model objects.}

\item{treatment}{Treatment name.}
}
\value{
A ggplot object.
}
\description{
Compares observed and predicted treatment means
across multiple fitted models.
}
