% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_models.R
\name{rank_models}
\alias{rank_models}
\title{Rank models}
\usage{
rank_models(comparison)
}
\arguments{
\item{comparison}{Output of compare_models().}
}
\value{
Ranked comparison table.
}
\description{
Ranks models based on multiple criteria.
}
