% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ankom.R
\name{read_ankom}
\alias{read_ankom}
\title{Import ANKOM RF output}
\usage{
read_ankom(file)
}
\arguments{
\item{file}{Path to ANKOM Excel file.}
}
\value{
A data frame containing raw ANKOM data.
}
\description{
Reads a raw ANKOM RF export file.
}
