% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{read_metadata}
\alias{read_metadata}
\title{Import metadata}
\usage{
read_metadata(file, sheet = "metadata")
}
\arguments{
\item{file}{Path to metadata file.}

\item{sheet}{Sheet name.}
}
\value{
Metadata table.
}
\description{
Reads ANKOM metadata.
}
