% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{rumenGP-package}
\alias{rumenGP}
\alias{rumenGP-package}
\title{rumenGP}
\description{
Tools for rumen gas production modeling.
}
\author{
\strong{Maintainer}: Arlan Araujo Rodrigues \email{arlan.rodrigues@wisc.edu}

Authors:
\itemize{
  \item Hilario Cuquetto Mantovani
}

}
\keyword{internal}
