% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{rumenGP-package}
\alias{rumenGP}
\alias{rumenGP-package}
\title{rumenGP}
\description{
Tools for rumen gas production modeling.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://araujorodrig-lab.github.io/rumenGP}
  \item \url{https://github.com/araujorodrig-lab/rumenGP}
  \item Report bugs at \url{https://github.com/araujorodrig-lab/rumenGP/issues}
}

}
\author{
\strong{Maintainer}: Arlan Araujo Rodrigues \email{araujorodrig@wisc.edu}

Authors:
\itemize{
  \item Hilario Cuquetto Mantovani \email{hcmantovani@wisc.edu}
}

}
\keyword{internal}
