% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.brody_fit.R
\name{summary.brody_fit}
\alias{summary.brody_fit}
\title{Summary of Brody fits}
\usage{
\method{summary}{brody_fit}(object, ...)
}
\arguments{
\item{object}{A brody_fit object.}

\item{...}{Additional arguments.}
}
\description{
Summary of Brody fits
}
