% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.custom_fit.R
\name{summary.custom_fit}
\alias{summary.custom_fit}
\title{Summary of Custom Model Fits}
\usage{
\method{summary}{custom_fit}(object, ...)
}
\arguments{
\item{object}{A \code{custom_fit} object.}

\item{...}{Not used.}
}
\value{
Invisibly returns the input
\code{custom_fit} object.
}
\description{
Summarizes a fitted custom nonlinear model.
}
\details{
The summary typically reports:

\itemize{
\item Model name
\item Model formula
\item Parameter estimates
\item Residual Sum of Squares (RSS)
\item Root Mean Squared Error (RMSE)
\item R-squared (R²)
\item Akaike Information Criterion (AIC)
\item Bayesian Information Criterion (BIC)
}

This method provides a concise overview of
parameter estimates and model performance for
user-defined nonlinear equations fitted with
\code{fit_custom()}.
}
\examples{

files <- example_data()

raw_data <- read_ankom(
  files$ankom
)

metadata <- read_metadata(
  files$metadata
)

gp <- process_ankom(
  raw_data,
  metadata,
  headspace_ml = 210,
  temperature_c = 39
)

custom_fit <- fit_custom(
  data = gp,
  formula =
    Gas_mL ~
      A *
      (
        Time_h /
        (
          Time_h + K
        )
      ),
  start = list(
    A = 150,
    K = 10
  ),
  lower = c(
    A = 0,
    K = 0
  ),
  model_name = "Hyperbolic"
)

summary(
  custom_fit
)

}
\seealso{
\code{\link{fit_custom}},
\code{\link{plot_fit}},
\code{\link{plot_residuals}},
\code{\link{compare_models}}
}
