% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.exp0_fit.R
\name{summary.exp0_fit}
\alias{summary.exp0_fit}
\title{Summary of EXP0 fits}
\usage{
\method{summary}{exp0_fit}(object, ...)
}
\arguments{
\item{object}{An exp0_fit object.}

\item{...}{Additional arguments.}
}
\description{
Summary of EXP0 fits
}
