% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.expl_fit.R
\name{summary.expl_fit}
\alias{summary.expl_fit}
\title{Summary of EXPL fits}
\usage{
\method{summary}{expl_fit}(object, ...)
}
\arguments{
\item{object}{An expl_fit object.}

\item{...}{Additional arguments.}
}
\description{
Summary of EXPL fits
}
