% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.lel_fit.R
\name{summary.lel_fit}
\alias{summary.lel_fit}
\title{Summary of LEL fits}
\usage{
\method{summary}{lel_fit}(object, ...)
}
\arguments{
\item{object}{A lel_fit object.}

\item{...}{Additional arguments.}
}
\description{
Summary of LEL fits
}
