% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mitscherlich_fit.R
\name{summary.mitscherlich_fit}
\alias{summary.mitscherlich_fit}
\title{Summary of Mitscherlich fits}
\usage{
\method{summary}{mitscherlich_fit}(object, ...)
}
\arguments{
\item{object}{A mitscherlich_fit object.}

\item{...}{Additional arguments.}
}
\description{
Summary of Mitscherlich fits
}
