% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mm_fit.R
\name{summary.mm_fit}
\alias{summary.mm_fit}
\title{Summary of Michaelis-Menten fits}
\usage{
\method{summary}{mm_fit}(object, ...)
}
\arguments{
\item{object}{A mm_fit object.}

\item{...}{Additional arguments.}
}
\description{
Summarizes Michaelis-Menten model fitting results.
}
