% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.orskov_fit.R
\name{summary.orskov_fit}
\alias{summary.orskov_fit}
\title{Summary of Ørskov and McDonald fits}
\usage{
\method{summary}{orskov_fit}(object, ...)
}
\arguments{
\item{object}{An orskov_fit object.}

\item{...}{Additional arguments.}
}
\description{
Summary of Ørskov and McDonald fits
}
