% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_ankom.R
\name{validate_ankom}
\alias{validate_ankom}
\title{Validate processed rumen gas production data}
\usage{
validate_ankom(data)
}
\arguments{
\item{data}{A rumen_gp object.}
}
\value{
The validated rumen_gp object.
}
\description{
Performs quality-control checks on a rumen_gp object.
}
\details{
Supports datasets created by either:
\itemize{
\item process_ankom()
\item as_rumen_gp()
}

ANKOM-specific checks are performed only when
Gas_PSI is available.
}
